Created parent class for phone menus
